Do not hard-code path to nologin executable
One of the tasks creates a user with a shell of `/bin/nologin`. This is
problematic, because the `nologin` executable may not be at that path.
For example, on Fedora 26 and 27, it's present at `/usr/sbin/nologin`.
On Arch Linux, it's available at `/usr/bin/nologin`.

Do not hard-code the path to the `nologin` executable. Discover the path
to this executable at runtime instead.