Drop pulp_web in favor of pulp_wsgi
The `pulp_web.service` starts the Django development web server. While
this is great for development, it's horrid for real deployments. The
development web server has terrible security and performance
characteristics.

Drop the Django development web server in favor of Gunicorn, a WSGI
server. This new server only does half of what the Django development
web server did; this new server runs the Pulp python code, but doesn't
listen for any HTTP connections. A separate commit will need to be
created in which web server support is brought in.