Prepend all roles with "pulp3"
The "webserver" role isn't meant to configure web servers in general.
Instead, it's meant to configure a web server for use with Pulp 3.
Similar, the "redis" role isn't meant to configure Redis in general.
Instead, it's meant to configure Redis for use with Pulp 3. However,
role names like "webserver" and "redis" give the misleading impression
that these roles are highly generic and can be used in many situations
that don't involve Pulp 3. This is incorrect.

Prepend these role names with "pulp3," so that users get the correct
impression.