Create pulp3-workers role
When executed, this role will install and configure several pulp worker
services, and will start, stop, enable and disable the services as
described by the `pulp3_workers` variable. (This variable has a default
value.)

Drop the call to the `pulp3-redis` role from the sample playbook,
because the new `pulp3-workers` role depends on it and automatically
pulls it in.

Fix: https://pulp.plan.io/issues/3731