Drop dashes from the top of a yaml file
Three dashes indicate the start of a YAML document. Anecdotally, they're
optional. Drop them from the top of a YAML file in the `pulp3-workers`
role, so that all YAML files follow the same format. The dashes could
also be added to many files, but dropping these dashes is lower impact
(in terms of number of files touched) and leads to more concise
documents.