Use systemd module
Use the systemd module instead of the service module for all service
management. Do this because the "systemd" module lets us do things with
the systemd daemon that the more generic "service" module. For example,
it lets us reload the systemd daemon right before changing the state of
services, which in turn prevents certain subtle bugs.

All of the platforms we're targeting use systemd.