Enable yamllint
Letting a linter tell us what the proper formatting for code should be
eliminates a type of bike-shedding.

yamllint and ansible-review have different opinions about proper code
formatting. But yamllint is a much easier tool to use, and it's
available to us *right now*. We can figure out how to re-enable
ansible-review later, if desired.