Switch to geerlingguy.postgresql role
The previous ANXS role made limited assumptions about how and where
to get Postgres. Further, the last commit on the git repo was 6 months
ago at the time of this commit. The idea is to move to a more active
role that support for using SCL based postgres can be added to.

This currently points at a fork to get the Fedora and SCL support.