Remove vagrant assets and restructure
Previously, the example playbooks (and docs) were coupled with vagrant
usage. This PR removes the vagrant assets, and revises the examples and
documentation to be applicable for those using ansible directly (without
Vagrant). As a bonus, the example playbooks and variables are symlinked
to molecule, so the examples are tested by travis for PRs against this
repository.

https://pulp.plan.io/issues/4325
re: #4325