Fix multiple breaks
1. Create a user for molecule in the "source" scenario. This fix isn't
ideal-- we need to make more adjustments to the users generally, but
there were multiple problems with the direct approach. Ideally for a
dev install, the developer could simply set the `pulp_user` to their
user, which is expected to exist on the managed node. However this
approach broke the pulplift installation. I'm sure it will work, we just
need to try again when everything else is already working.

2. Minor fix: incorrect var name in pulp3-devel. Failure was hidden by
the first test failure.

3. Move collectstatic to handler. Multiple events cause static files to
be collected, so it makes more sense for this to be a handler. This
fixes idempotence tests, which again were hidden by the previous errors.
This was only an issue for installations that use the pulp3-devel role.
[noissue]