Add pulp_content_bind variable for pulp-content-app.service
On /etc/pulp/settings.py we need to point which CONTENT_HOST is
being used by pointing to `localhost:8080` ot `fqdn:8080` this
is being added by `pulp_content_host` variable.

This same value is being used to render the `pulp-content-app.service.j2`
on its `--bind` property.

Problem:

Gunicorn fails with `example.myhost.com:8080` because it will be invalid
setting to the default `localhost` is also not a good idea when we need
to consume content from other hosts.

Solution:

Add a new variable `pulp_content_bind` which defaults to `0.0.0.0:8080`
thus it enabled access from any host/interface unless specified via
 `-e pulp_content_bind=...`

Having different variables for each of those settings is the best to do.

[noissue]