Allow bind and nginx to be different
This introduces a pulp_api_bind variable in the 'pulp' playbook which
the user can use to set the bind address (or multiple) of pulp-api. This
is used by the pulp-api service template for the systemd service.

It moves the nginx and apache reverse proxy variables out of the 'pulp'
playbook and into the 'pulp-webserver' playbook where they are used
exclusively. The defaults are there also. It created 4 variables:

* pulp_content_host
* pulp_content_port
* pulp_api_host
* pulp_api_port

https://pulp.plan.io/issues/4964
closes #4964