Problem: Pulp 2 stops working after installing Pulp 3
Solution: make use of the 'pulp' group when setting permissions

Pulp 2.20 changed ownership of /var/lib/pulp and /etc/pulp from group 'apache' to group
'pulp'. This change to the installer ensures that the permissions on these directories
remain the same after the installer runs on a machine that already has Pulp 2 installed.

This patch introduces 3 new pulp settings called 'pulp_group', 'pulp_group_id', and
'pulp_user_id'. The 'pulp_group' defaults to 'pulp'. The 'pulp_user' is added to the
'pulp_ group' group. The connection is reset after this to support installations where
the ssh user being used for ansible is the same as the pulp_user. This allows the user
changes to take affect. The defaults for 'pulp_user_id' and 'pulp_group_id' are empty.
When specified, the installer will use values to set 'uid' and 'gid' respectively.

closes: #5104
https://pulp.plan.io/issues/5104