Add PostgreSQL 9.6 for CentOS7
Pulp is requiring PostgreSQL 9.6 with Django 2.2 (LTS). This upgrades
the version of PostgreSQL on CentOS 7 to 9.6 via software collections.

The systemd files can no longer depend on the postgresql service name to
restart them when postgresql restarts. Foremost because we can't be
guaranteed postgresql is a systemd unit on this system. Secondly due to
this PR that name is now changing on centos7.

This PR removes the systemd dependency on postgresql service units and
instead configures all Pulp services to automatically always restart
when it exits. Stopping the unit with systemctl still stops it.

https://pulp.plan.io/issues/5147
closes #5147