Install JWT authentication plugin for HTTPie
The plugin is required for querying the pulp content server
with Json Web Token authorization headers. HTTPie uses a
configuration provided by .netrc by default. A custom
Authorization header is always overwritten by this
configuration.

This change is relevant to pulp_docker plugin, where an
authenticated registry for a content pulling/pushing is
going to be implemented, and HTTPie is the recommended
tool to perform API requests.

[noissue]