Adds a pulpcore/plugin version compatibility check
This change also reorders the running of pre-requisite roles and the creation of the virtualenv.
The version compatibility check is performed before any other packages are installed into the
virtualenv.

A requirement.in file is generated using the pulp_version and pulp_install_plugins settings passed
in from the playbook. pip-compile inspects all the packages in the requirements.in and returns
any conflicts between version requirements.

Some plugins such as pulp_rpm have to have build dependencies installed before the setup.py can
be fully interpreted by pip-compile. An error raised in such a case is ignored because it occurs
after the versions of direct dependencies of the plugin have been evaluated.

fixes: #6189
https://pulp.plan.io/issues/6189