Problem: pre-flight check does not work properly when it needs a prereq role
Solution: Run it only once and allow no errors if there are 0 prereq roles.
Run it 2x and allow errors (but not failures) only on the 1st time
if there are prereq roles.

fixes: #6645
pre-flight check does not work properly when it needs a prereq role
https://pulp.plan.io/issues/6645

re: #6623
pulpcore and plugin pre-flight check seems to not be enforcing and then installer fails at collectstatic
https://pulp.plan.io/issues/6623