Problem: pre-flight check does not check plugins that are installed but not listed in pulp_install_plugins
Solution: List them in requirements.in
Implementation includes:
1. New non-user-facing variable pulp_install_plugins_normalized with
plugin names corrected from underscores to dashes.
2. Having the entire repo use pulp_install_plugins_normalized
(a safety measure so that the same values used for pre-flight checks
is used for installation.)
3. A currently unadvertised variable pulp_irregularly_named_plugins
containing a list of pip packages that are plugins but do not begin
with "pulp-".
4. Print pulp_install_plugins variables to facilitate debugging,
only if verbosity level is at least 1 (-v).

fixes: #6644
pre-flight check does not check plugins that are installed but not listed in pulp_install_plugins
https://pulp.plan.io/issues/6644

re: #6623
pulpcore and plugin pre-flight check seems to not be enforcing and then installer fails at collectstatic
https://pulp.plan.io/issues/6623