Problem: pre-flight check errors when a package is installed
system-wide at a version that is not available on PyPI, and

This happens for old pulp_rpm prereqs.

This is a regression implemented in the referenced issue's fix.

Solution: Have `pip freeze` not list system-site packages.
`pip-compile` will fail if compatible packages are not available
on PyPI. However, they are, and Pulp does not rely on system-site
packages anymore.

This makes the check behave like before the referenced issue's fix.

fixes: #6689
pre-flight check errors when a package is installed system-wide at a version that is not available on PyPI
https://pulp.plan.io/issues/6689

re: #6642
pre-flight check does not analyze older pulp 3 installs properly
https://pulp.plan.io/issues/6642