Problem: pre-flight check errors when there is an upgraded pip
in a system-wide venv

and when the package is installed as an RPM in the prereq role
at an acceptable version, so the pre-flight check should not
have tried to use a newer version from PyPI.

This is a regression from #6645. Previously many errors were ignored,
this error should still be ignored too until system-wide packages
support is removed.

fixes: #6690
pre-flight check errors when there is an upgraded pip in a system-wide venv
https://pulp.plan.io/issues/6690

re: #6645
pre-flight check does not work properly when it needs a prereq role
https://pulp.plan.io/issues/6645