LD_LIBRARY_PATH: Allow one to specify a specific path
When using SCL, per its pattern, the LD_LIBRARY_PATH needs to be
specified in order for the libraries/binariess to find the "right"
version.

The scripts dropped in `/etc/profile.d` are working for login shells
but tasks we are running are using the pulp user which has no login
shell (/sbin/nologin).

Given `rh-postgresql96` is currently used (and `postgresql-devel-9.2`)
gets installed during installation the lack of explicitly specifying it
might have gone un-noticed.

But when dealing with replacing `rh-postgresql96` by `rh-postgresql10`
(and all related files) or simply installing the
`rh-postgresql10-postgresql-libs` and `rh-postgresql10-postgresql-devel`
to connect to a PG10 server setup else where it fails with the following
errors:

 * django.db.utils.OperationalError: SCRAM authentication requires libpq
   version 10 or above

 * /opt/rh/rh-postgresql10/root/bin/initdb: error while loading shared
   libraries: libpq.so.rh-postgresql10-5: cannot open shared object
   file: No such file or directory

fixes #6913