Do not call packages module iteratively
This improves performance on one task, cleans up code for
another.

https://github.com/oasis-roles/meta_standards/blob/1.0.0/README.md
Avoid calling the package module iteratively with the {{ item }}
argument, as this is impressively more slow than calling it with
the line name: "{{ foo_packages }}". The same can go for many
other modules that can be given an entire list of items all at once.

[noissue]