pulp_devel: Implement support for signing service script
being copied & registered.

This is in pulp_devel because the signing service is not in
its permanent packaging. This is a temporary installation
method.

fixes: #7247
https://pulp.plan.io/issues/7247
As a pulp_installer developer-user, the pulp_rpm signing service will be installed for me

Related to 4812
https://pulp.plan.io/issues/4812
As a user, I can publish a Yum repository that works with repo_gpgcheck=1 (Signed Repositories)