Adopt a new pattern to specify defined/undefined for tls variables
The goal of this commit is to adopt a new pattern for TLS related
variables so it can be easily embedded within another playbook.

Current Issue:

- name: Install Automation Hub via Pulp Installer
  include_role:
    name: "{{ item  }}"
  vars:
    pulp_webserver_tls_cert: "{{ mycert | default('') }}"
    pulp_webserver_tls_key: "{{ mykey | default('') }}"
  with_items:
    - "pulp.pulp_installer.pulp_redis"
    - "pulp.pulp_installer.pulp_services"
    - "pulp.pulp_installer.pulp_webserver"

That would be an ideal approach, but this would fail currently because
the variable itself - within the pulp webserver role scope - ends up
being defined (but empty) - so the play ends with src (or content) is
required.

With the approach suggested in this PR - This would easily allow people
to have a layer on top of the Pulp installer, and if they wish to
specify the value they can, else by default it would be considered
"undefine"

[noissue]