Allow one to inherits from another package
The scenario is the following. I provide to the public a plugin that
pulls another plugin and add some feature on top of it. Both my plugin
and the plugin I pull conflicts (from a package definition standpoint)
so I can declare just one of them. Yet I'd like to leverage the content
of my inherited plugin like snippets.

This would work like the following:

```
pulp_install_plugins:
  automation-hub:
    inherits: galaxy-ng
```

This would led to `python3-automation-hub` plugin to be installed
(as expected) and also apply the snippet from `galaxy-ng` plugin which
my plugin rely upon,

[noissue]