Disable fips jobs in PR and release builds
The pulp org is limited to 20 concurrent jobs in Github Actions and
we're starting to see queue times of 15+ minutes. These jobs get kicked
off every time a change is pushed to a PR. I recommend disabling them
for PRs and releases. We run these jobs nightly and I think that'll
suffice.

[noissue]