Add ability to specify a minor version for `pulpcore_version`
`pulpcore_version` now defaults to a specific minor release of pulpcore (e.g.: 3.15)
which the installer is going to install.

When the installer is run for the first time, the latest patch release of pulpcore
(e.g.: 3.15.2) is installed. If the installer is run again with the same value of
`pulpcore_version`, the pulpcore package is not updated. In order to update to the
latest patch release within a minor stream, the user must set `pulpcore_update` to
true.

If the user sets the `pulpcore_version` to a different minor release of pulpcore
(e.g.: 3.16), the installer will attempt to upgrade pulpcore to the latest patch
release (e.g.: 3.16.1) for that minor release of pulpcore.

If the user sets the `pulpcore_version` to a specific patch release, the installer
will attempt to install that specific release of pulpcore.

If the installed plugins are not compatible with the requested release of pulpcore,
the installer will not make any changes.

The `version` of a plugin can now also be specified as a minor release (e.g.: 1.8) or
a specific patch release (e.g.: 1.8.2). When the a minor release is specified and it
is the same as the installed version of the plugin, the `upgrade` parameter for that
plugin must be set to `true` in order for the plugin to be updated to the latest
patch release.

The internal variable `__pulp_version` was renamed to `__pulpcore_version`.

The release-static test scenario for molecule was also updated to test this feature.

fixes: #8847
https://pulp.plan.io/issues/8847