Problem: pulp_installer 3.7 is failing to upgrade the dynaconf RPM with the external repo
Workaround: python3-dynaconf to the undeclared dependencies list.
This makes `pulp_pkg_upgrade_all: true` upgrade it, users can workaround
it that way.

We are doing this workarond because it is a bug in the RPM packaging
that is fixed on newer pulpcore repos than 3.7.

Actually fixes: #9181

(But this is part of a larger PR, so we will have to manually close the
bug)
[noissue]

(cherry picked from commit 1d374e3d2de5695ca0273bad669afdbd56d69a91)