Fix SELinux labels not being not set on files
that are created by the installer.

There are 3 ways this is done:
1. serole/seuser/setype on files created before the handler
"Load the SELinux policy packages" gets run -
these are for when the installer is re-run, to correct drift.
2. serole/seuser/setype on files created after said handler -
these apply the actual selinux labels to the files.
3. The 2 handlers to restore contexts - these are needed for apply
the context to files created before the "Load" handler gets run,
and for recursion.

Note that in the future, we may replace the handlers that call commands
with `file` module tasks or handlers that recurse.

fixes: #962