Replace molecule inspec tests with testinfra
I had to pass the python path to molecule because each combination of
scenario + platform was using a different path for it.
I believe that this kind of issue didn't happen with inspec because
inspec does not use python.

Another thing that I had to verify and configure for each scenario/platform
was the pulp-smash that was causing python exceptions (config file not found)
in some situations.

fixes: #1009